fix: Hint labels showing default text regardless of localization
Motivation:

Regardless of the presence of localized label for the hint button, when
toggling the button, default hardcoded text would be displayed.

Modification:

In the `toggleLabel` function in `HintView.js`, `polyglot` is used to
translate the hint labels rather than using hardcoded texts.

Result:

The hint button correctly shows the localized labels for the hint
button and falling back to the English version if it is absent.

Fixes #1944